postinst: Make cleanup of local AppArmor abstraction conditional
Right now we perform the cleanup unconditionally, but there's
a scenario in which that could be problematic.

Old versions of the abstractions, including those shipped in
bookworm, were written with AppArmor 2 in mind, which didn't
implement the "include if exists" directive.

If we were to bring forward such a configuration, specifically
because of the abstractions undergoing the conffile transfer
process, and also deleted the local abstractions at the same
time because they're empty, the profiles would become invalid
and VM startup completely broken.

In order to avoid that possibility, only delete the local
abstractions if the AppArmor configuration has not been
customized by the local admin. In that case, we know that the
abstraction itself will use the "include if exist" directive
and thus that it's fine for the local abstractions not to be
present.

Gbp-Dch: Ignore