control: Introduce common package
The libvirt.conf and libvirt-admin.conf configuration files
are parsed by the library itself to ensure that whatever client
is used, including third-party ones such as virt-manager, will
obey the settings contained therein.

Accordingly, the files should live in the library package, not
together with the default clients.

However, since the library package itself is Multi-Arch aware
and so it's possible to install multiple copies of it in
parallel, it's likely not a good idea to include conffiles in
it.

Take a slightly different approach, and create a common package
for the conffiles to live in. Since we're at it, we can move
all existing data files to it as well, eliminating the risk to
have any conflict whatsoever.

All other binary package currently point to libvirt0 for their
docs, and we have to switch them to libvirt-common instead.
Luckily the dir_to_symlink dpkg-maintscript-helper command is
very robust, and will behave correctly if the switch to symlink
has already been performed in a previous update: in that case,
the old symlink will simply be replaced with the new one.

Gbp-Dch: Ignore