control: Introduce ARCHES_DMIDECODE
This was made a Recommends to cope with the fact that dmidecode
is not available on all architectures. A better solution is to
make it a Depends only on the architectures where libvirt would
actually use it, which as it happens are also the ones where the
package gets built.