control: Drop Recommends on init system
This is a leftover from when we had separate daemon-system-systemd
and daemon-system-sysv packages and we needed to ensure that only
one of them, specifically the one matching the host's init system,
would be installed.

Now that we are back to shipping both systemd unit files and sysv
init scripts in a single package, this dependency has become
pointless. Not explicitly depending on an init system, in addition
to being more in line with the usual practice for daemons in
Debian, has the advantage of making them suitable for running
inside a container, where the presence of an init system is
generally not necessary or desirable.

Unfortunately this doesn't quite solve things for us since we
still depend on polkitd, which depends on logind, which in turn
will bring in the init system, but at least it gets us closer.