control: Drop Depends on firewalld
libvirt needs to integrate with firewalld to ensure that the
packet filtering rules it creates to implement guest networking
work as expected and don't get in the way, but if firewalld is
not present on the system there is no loss of functionality.

Hence the dependency is not justified, even as the non-default
option.