patches: Drop forward/Reduce-udevadm-settle-timeout-[...]
This was added in 2012 to work around an issue (#663931) where
running pool-refresh took too long, causing virt-manager to
fail connecting because of a timeout.

I'm assuming that things have improved in the 10+ years since,
both on the kernel and udev side, and that this workaround is
no longer needed at this point. As far as I can tell, no other
distro is carrying a similar patch either.

If we get further reports in the future, indicating that this
is indeed still a problem, we should look into a proper
upstream fix instead of a Debian-only workaround.