control: Once again prefer iptables for the network driver
The nftables backend was unfortunately found to have serious
compatibility issues with popular software such as Docker and
ufw (the default firewall in Ubuntu). Given the situation, it
doesn't seem wise to make it the default backend in Debian quite
yet.

This reverts commit 4a7c66d5c3cfaf5247518cd6d321424672803f3d
almost completely; however, the possibility to have nftables
satisfy the dependencies for the network driver is retained.
This will allow users to manually switch backends without
needing to have iptables installed as well.

Closes: #1090355