daemon: Drop Recommends on non-QEMU hypervisor drivers
The various hypervisor drivers were moved to their own
packages way back in 2018, and at that time they were kept
as Recommends so that users would not be hit by unexpected
functionality loss after upgrading.

At this point, anyone upgrading will already have the
specific packages installed, so these Recommends no longer
serve any useful purpose; on the contrary, they can result
in undesired hypervisor drivers getting automatically
dragged in during upgrade or initial installation.

We retain the Recommends for the QEMU driver for now,
because we know that there are packages in Debian which
need the QEMU driver and don't depend on it explicitly,
instead relying on the daemon package dragging it in.
That's something that we should sort out eventually, but
for now let's avoid the potential disruption.