control: Introduce ARCHES_WIRESHARK
Wireshark hasn't been built successfully on either m68k or
x32 since 2023, and that in turn prevents libvirt from being
built on those architectures.

Since the Wireshark dissector is an optional feature, we can
simply disable it. This should make libvirt buildable on both
m68k and x32.

If the situation with Wireshark changes in the future,
enabling the feature again on those architectures once again
will be easy enough.