changelog: Import entry for 11.3.0-2
11.3.0-2 was prepared on the debian/trixie branch but
in retrospect that was premature, since the upload ended
up still going through unstable.

We want to retain the changelog entry, specifically due
to the mention of #1104735 being fixed, but merging the
branch into debian/latest at this point would just
complicate the git history for no benefit. Let's just
import the changelog entry and move on.

Gbp-Dch: Ignore