Tweak details for libvirt-qemu user
Use proper capitalization in the GECOS field and make it
more descriptive; use /nonexistent as home directory, since
it's not supposed to be used for logging in anyway; explicitly
ask for the libvirt-qemu user to have libvirt-qemu as the
primary group; use "-" (empty) as the value for the login
shell field instead of not listing it.

This only impacts new installations: existing users will not
be modified.